- started new development branch

+++ examples/familiestopersons/BenchmarxFamiliesToPersons/src/org/benchmarx/examples/familiestopersons/implementations/medini/MediniQVTFamiliesToPersons.java
@@ -131,8 +131,11 @@ public class MediniQVTFamiliesToPersons extends BXToolForEMF<FamilyRegister, Per
 		// delete content of traces folder
 		File tracesFolder = new File("./src/org/benchmarx/examples/familiestopersons/implementations/medini/base/traces");
 		final File[] files = tracesFolder.listFiles();
-		for (File f : files) {
-			f.delete();
+		if (files != null) {
+			for (File f : files) {
+				if (f != null)
+					f.delete();
+			}
 		}
 		
 		// Initialise resource set of models
